fix issue with left over rendering artifacts from info popup
turn the tags/keywords into `Span`s in the info popup so that
when the underlying content is re-rendered, the text is properly
removed. For some reason the way Ratatui was doing the rendering
caused a Line made from plain String with spaces wasn't cleaned
up correctly. Not sure if this is the best/right way but at least
it's rendered correctly now - or at least no left over stuff shown.